refactor(test): extrair torgb para src/test/colors.ts
helper torgb() estava copy-pasteado em dois arquivos de teste.
extraído para utilitário compartilhado com normalização adicional
de espaços em strings rgb() variacionais.
